@@ -25,7 +25,6 @@ Step 4:
 import pathlib
 import sys
 import xml.etree.ElementTree
-from typing import List  # mypy!
 
 # -
 
@@ -146,7 +145,7 @@ db.load_xml(XML_DIR / "egl.xml")
 
 # -
 
-lines: List[str] = []  # noqa: E999 (bug 1573737)
+lines: list[str] = []
 
 keys = sorted(db.consts.keys())
 
@@ -166,7 +165,7 @@ for lib in db.LIBS:
 
 # -
 
-b_lines: List[bytes] = [HEADER] + [x.encode() for x in lines] + [FOOTER]
+b_lines: list[bytes] = [HEADER] + [x.encode() for x in lines] + [FOOTER]
 b_data: bytes = b"\n".join(b_lines)
 
 dest = pathlib.Path("GLConsts.h")
