@@ -18,7 +18,6 @@ import sysconfig
 import tempfile
 import warnings
 from contextlib import contextmanager
-from functools import lru_cache
 from pathlib import Path
 from typing import Callable, Optional
 
@@ -41,7 +40,7 @@ PIP_NETWORK_INSTALL_RESTRICTED_VIRTUALENVS = ("mach", "build", "common")
 _is_windows = sys.platform == "cygwin" or (sys.platform == "win32" and os.sep == "\\")
 
 
-@lru_cache(maxsize=None)
+@functools.cache
 def use_uv():
     return (
         os.environ.get("MACH_NO_UV", "").lower()
@@ -53,12 +52,12 @@ def use_uv():
     )
 
 
-@lru_cache(maxsize=None)
+@functools.cache
 def get_uv_executable():
     return shutil.which("uv")
 
 
-@lru_cache(maxsize=None)
+@functools.cache
 def show_pip_output():
     return os.environ.get("MACH_SHOW_PIP_OUTPUT", "").lower() in ("1", "true")
 
@@ -1169,7 +1168,7 @@ class ExternalPythonSite:
         self._prefix = os.path.dirname(os.path.dirname(python_executable))
         self.python_path = python_executable
 
-    @functools.lru_cache(maxsize=None)
+    @functools.cache
     def sys_path(self):
         """Return lists of sys.path entries: one for standard library, one for the site
 
@@ -1241,7 +1240,7 @@ class ExternalPythonSite:
         return stdlib
 
 
-@functools.lru_cache(maxsize=None)
+@functools.cache
 def resolve_requirements(topsrcdir, site_name):
     thunderbird_dir = Path(topsrcdir, "comm")
     is_thunderbird = thunderbird_dir.exists() and any(thunderbird_dir.iterdir())
